[PYG-356] ✍🏼 Typo in helper text (#448)
## Reported from user

![image](https://github.com/user-attachments/assets/4ea75c74-2288-4a43-9241-b6bca432bca3)

## Description
The current code assumes that the views are using `PascalCase` in CDF.
The fix looks up what is available in the module and the helper message
based on that.

Note that I do not trigger a new release for this one as it is no change
in behavior, only a helper text.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip