[PYG-352] 🫡 Support excluding views from generated SDK (#450)
# Description

The motivating use case for this is using Pygen in streamlit
applications inside of CDF. This is currently very slow for SDKs that
are generated. This is especially prevalent for CogniteCore extensions
that have 33 extra views that are typically not used.

## Bump

- [ ] Patch
- [x] Minor
- [ ] Skip

## Changelog
### Added

- New parameters `exclude_views` and `exclude_spaces` to the
`generate_sdk` and `generate_sdk_notebook` functions. This is to enable
using a subset of the data models when generating an SDK.