[PYG-364] 🌴 Pagination cursor part 1 (#460)
# Description

**REVIEWER**: This is a +100 line code change (50 lines changed + 50
line tests) the remaining 300 lines are generated (everything inside
/examples). The last commit contains the generated code.

This is part 1 of preparing for returning and accepting cursor as an
argument for the iterate method. It is needed as each batch
call/iteratoin must return a set of cursors that can be used for the
next iterations. That cannot be done when the chunking is done on the
client side as it is today.

Found 2 small unrelated bugs in the refactoring, see comments in the
code.

## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Changed

- In the `.iterate()` method, Pygen now uses the `chunk_size` to set the
max limit in the request size. For example, if `chunk_size=100` and
`limit=1000`, then Pygen will now do 10 requests with size 100 instead
of one with 1000 and chunk it on the client side.

### Fixed

- Pygen will now print a warning if it reduces the batch size when
executing large queries. This can happen in the generated
`.list/select/retrieve/iterate()` methods.