[PYG-404] 🐛Fix skipping reverse direct relations that points to views in the data model. (#497)
# Description

The user got the following error message:
<img width="1106" height="317" alt="image"
src="https://github.com/user-attachments/assets/d52cce03-fdbf-42e5-9eb2-8536e69c7dc4"
/>

Even though `tags.children` and `tags.timeSeries` are reverse direct
relations pointing to views that are in the model.


## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Fixed

- When generating an SDK, `pygen` no longer skips reverse direct
relations that has been overwritten in a view that implements parent
view and the parent view is not part of the data model.
- When generating an SDK, `pygen` uses the `source` for reverse direct
relations to determine type if it is available and falls back on the
through parameter.