[PYG-412] 🧑‍🔧 Extend omni model (#503)
# Description

**Reviewer**: All code inside `examples/` is generated. This PR thus
contains 0 lines of Python, only 44 lines of YAML with the new view
definition.

**Context**: `Omni` is a data model designed to test `pygen`. It covers
all possible value types, connections, as well as different types of
uses of implements.

This PR extends this model with a new view `ConnectionItemH`. This has a
direct relations to a parent view. The intention of this is to be able
to test that when calling the `.as_write()` method that take a response
DTO (data transfer object) and converts it to a request DTO, then if the
direct relation is a child view that information should persist. Note
this PR is only extending the model, the bugfix will come in later.

<img width="1526" height="605" alt="image"
src="https://github.com/user-attachments/assets/497bad45-b2ed-4f07-a020-1fcf59e8f40f"
/>


## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip