[🐛] Fx switch to uv backend (#537)
# Description

The issue is that `hatchling` build the source distribution with the
entire repo, the wheel is built correctly.

This causes the source distribution to become large which will make us
hit the 10GB limit of `pypi.org` for the package.

Switching to uv is faster and builds both the wheel and source
correctly.

<img width="1269" height="959" alt="image"
src="https://github.com/user-attachments/assets/76accfd1-4690-4d1b-b92a-43876599b2f5"
/>

## Bump

- [x Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Improved

- Changed backend for building the Toolkit package from `hatchling` to
`uv_build`.